+++ server/src/controllers/user.controller.js
@@ -1,3 +1,4 @@
+const mongoose = require("mongoose");
 const User = require('../models/User');
 
 const userController = {
@@ -12,6 +13,10 @@ const userController = {
 
     getUserById : async (req, res) => {
         try {
+            if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+                return res.status(400).json({ message: "ID tidak valid" });
+            }
+
             const user = await User.findById(req.params.id);
             if (!user) return res.status(404).json({ message: 'User not found' });
             res.json(user);
@@ -32,6 +37,10 @@ const userController = {
 
     updateUser : async (req, res) => {
         try {
+            if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+                return res.status(400).json({ message: "ID tidak valid" });
+            }
+
             const updatedUser = await User.findByIdAndUpdate(
                 req.params.id,
                 req.body,
@@ -46,6 +55,10 @@ const userController = {
 
     deleteUser : async (req, res) => {
         try {
+            if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+                return res.status(400).json({ message: "ID tidak valid" });
+            }
+            
             const deletedUser = await User.findByIdAndDelete(req.params.id);
             if (!deletedUser) return res.status(404).json({ message: 'User not found' });
             res.json({ message: 'User deleted' });
